Rename User to AdminUser

+++ db/migrate/20131222103406_remove_active_admin_comments.rb
@@ -1,5 +1,6 @@
 class RemoveActiveAdminComments < ActiveRecord::Migration
   def change
+    return unless table_exists?(:active_admin_comments)
     drop_table(:active_admin_comments)
   end
 end
